Changed the blockquote keyboard shortcut to Ctrl-T.  This changes the tab stops in most word processors so I think I can use it here.  The mnemonic is t = text.
It is very helpful to not use Ctrl-. for the blockquote shortcut.  The way that keyboard events are handled assumes that the key code is equal to the char code, which is only true in alphanumeric characters.

I don't feel bad if some browser uses Ctrl-T for anything since we use normally bound keys in WMD anyway.  e.g. horizontal rules can be created with Ctrl-R, which is refresh in Firefox.  I just wanted to avoid the obvious Ctrl-Q since that closes the browser window without warning in Safari.  That seemed a little too dangerous.